requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain
python-dotenv
langchain-openai
langchainhub